fix: Input handler 수정
